add rd_kafka_set_thread_name()
This function allows setting the thread local used by rdkafka to hold
the thread's name as well as updating what the OS names the thread if
possible.  Currently setting the OS's name is only supported on Linux,
but it appears to be possible on Windows and some versions of Solaris.